cast all as int

+++ api/date.go
@@ -13,8 +13,8 @@ func Handler(w http.ResponseWriter, r *http.Request) {
 	totalMinutes := int64(diffTime.Minutes())
 	totalDays := int64(totalHours / 24)
 	totalWeeks := int64(totalDays / 7)
-	remainderHours := totalHours % 24
-	remainderDays := extradays % 7
+	remainderHours := int64(totalHours % 24)
+	remainderDays := int64(extradays % 7)
 	remainderMinutes := int64(totalMinutes - totalHours*60)
 
 	fmtTime := fmt.Sprintf("%d Weeks %d Days %d Hours %d Minutes", totalWeeks, remainderDays, remainderHours, remainderMinutes)
